Working on refactoring delivery to be a POVA (Plain Old Value Object) and having a DeliveryGenerator that does the logic to create the params for it
